Split ingest-starter module out of ingest-runner
