+++ java/cdk/src/main/java/sleeper/cdk/stack/IngestStack.java
@@ -266,7 +266,7 @@ public class IngestStack extends NestedStack {
         IBucket jarsBucket = Bucket.fromBucketArn(this,
                 "jarsBucket-ingest",
                 "arn:aws:s3:::" + instanceProperties.get(JARS_BUCKET));
-        Code code = Code.fromBucket(jarsBucket, "ingest-" + instanceProperties.get(VERSION) + ".jar");
+        Code code = Code.fromBucket(jarsBucket, "ingest-starter-" + instanceProperties.get(VERSION) + ".jar");
 
         // Run tasks function
         String functionName = Utils.truncateTo64Characters(String.join("-", "sleeper",
