+++ java/ingest/pom.xml
@@ -28,6 +28,7 @@
 
     <modules>
         <module>ingest-core</module>
+        <module>ingest-starter</module>
         <module>ingest-runner</module>
         <module>ingest-status-store</module>
     </modules>
